Make common.py
